feat: split file and add version command
