+++ cli/logs.py
@@ -1,0 +1,5 @@
+import click
+
+@click.group()
+def logs():
+  pass
